// apps/api/src/modules/reservation/reservation.status.ts
export enum ReservationStatus {
  PENDING_APPROVAL = "PENDING_APPROVAL",
  CONFIRMED = "CONFIRMED",
  DRIVER_ASSIGNED = "DRIVER_ASSIGNED",
  DRIVER_ACCEPTED = "DRIVER_ACCEPTED",
  DRIVER_EN_ROUTE = "DRIVER_EN_ROUTE",
  CUSTOMER_PICKED_UP = "CUSTOMER_PICKED_UP",
  IN_PROGRESS = "IN_PROGRESS",
  COMPLETED = "COMPLETED",
  CANCELLED = "CANCELLED",
  NO_SHOW = "NO_SHOW",
  DELAYED = "DELAYED",
  FAILED = "FAILED"
}